fix: resolve tax extension auth timing, filename sanitization, and migration types
- Move setUser() call into controller middleware so it runs after
  auth:api has authenticated the request (fixes unauthenticated user binding)
- Remove premature auth check from TaxServiceProvider; repository is
  now set up lazily via the controller middleware instead
- Sanitize tax profile name before using it in CSV export filename
  to prevent path traversal / invalid filename characters
- Change user_id, tax_profile_id, and tag_id from unsignedBigInteger
  to unsignedInteger in migration to match Firefly III users table FK type
- Add /env to .gitignore

+++ database/migrations/extensions/2026_03_21_000001_create_tax_tables.php
@@ -16,7 +16,7 @@ return new class extends Migration {
     {
         Schema::create('extension_tax_profiles', static function (Blueprint $table): void {
             $table->id();
-            $table->unsignedBigInteger('user_id');
+            $table->unsignedInteger('user_id');
             $table->string('name', 255);
             $table->unsignedSmallInteger('tax_year');
             $table->decimal('tax_rate', 5, 2)->default(0.00);
@@ -30,8 +30,8 @@ return new class extends Migration {
 
         Schema::create('extension_tax_tag_links', static function (Blueprint $table): void {
             $table->id();
-            $table->unsignedBigInteger('tax_profile_id');
-            $table->unsignedBigInteger('tag_id');
+            $table->unsignedInteger('tax_profile_id');
+            $table->unsignedInteger('tag_id');
             $table->timestamps();
 
             $table->foreign('tax_profile_id')->references('id')->on('extension_tax_profiles')->onDelete('cascade');
